requirements.txt
aioredis==1.3.1
amqp==5.1.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2023.5.7
cffi==1.15.1
channels==3.0.1
channels-redis==2.4.2
charset-normalizer==3.1.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.1
daphne==3.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.18
django-allauth==0.54.0
django-appconf==1.0.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-ckeditor==6.5.1
django-ckeditor-5==0.2.7
django-classy-tags==4.0.0
django-extensions==3.2.1
django-humanize==0.1.2
django-imagekit==4.1.0
django-js-asset==2.0.0
django-redis==5.2.0
django-social-share==2.3.0
django-taggit==4.0.0
django-taggit-templatetags==0.2.5
django-taggit-templatetags2==1.6.1
django-templatetag-sugar==1.0
django-timezone-field==4.1.2
djangorestframework==3.14.0
executing==1.2.0
exif==1.6.0
greenlet==2.0.2
hiredis==2.2.3
humanize==4.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.14.0
jedi==0.18.2
kombu==5.3.0
matplotlib-inline==0.1.6
msgpack==0.6.2
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pi==0.1.2
pickleshare==0.7.5
pilkit==2.0
Pillow==9.5.0
plum-py==0.8.6
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
redis==3.5.3
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0
